moved duplicated code to utils class
