+++ pacman/operations/routing_info_allocator_algorithms/malloc_based_routing_allocator/malloc_based_routing_info_allocator.py
@@ -14,8 +14,6 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 import logging
-import numpy
-from past.builtins import xrange
 from spinn_utilities.progress_bar import ProgressBar
 from spinn_utilities.log import FormatAdapter
 from pacman.model.graphs.common import EdgeTrafficType
@@ -26,8 +24,7 @@ from .key_field_generator import KeyFieldGenerator
 from pacman.model.routing_info import (
     RoutingInfo, BaseKeyAndMask, PartitionRoutingInfo)
 from pacman.utilities.utility_calls import (
-    check_algorithm_can_support_constraints,
-    compress_from_bit_array, expand_to_bit_array)
+    check_algorithm_can_support_constraints, get_key_ranges)
 from pacman.utilities.algorithm_utilities import ElementAllocatorAlgorithm
 from pacman.utilities.algorithm_utilities.routing_info_allocator_utilities \
     import (check_types_of_edge_constraint, get_edge_groups)
@@ -207,44 +204,6 @@ class MallocBasedRoutingInfoAllocator(ElementAllocatorAlgorithm):
         partition_info = PartitionRoutingInfo(keys_and_masks, group)
         routing_infos.add_partition_info(partition_info)
 
-    @staticmethod
-    def _get_key_ranges(key, mask):
-        """ Get a generator of base_key, n_keys pairs that represent ranges
-            allowed by the mask.
-
-        :param int key: The base key
-        :param int mask: The mask
-        :rtype: iterable(tuple(int,int))
-        """
-        unwrapped_mask = expand_to_bit_array(mask)
-        first_zeros = list()
-        remaining_zeros = list()
-        pos = len(unwrapped_mask) - 1
-
-        # Keep the indices of the first set of zeros
-        while pos >= 0 and unwrapped_mask[pos] == 0:
-            first_zeros.append(pos)
-            pos -= 1
-
-        # Find all the remaining zeros
-        while pos >= 0:
-            if unwrapped_mask[pos] == 0:
-                remaining_zeros.append(pos)
-            pos -= 1
-
-        # Loop over 2^len(remaining_zeros) to produce the base key,
-        # with n_keys being 2^len(first_zeros)
-        n_sets = 2 ** len(remaining_zeros)
-        n_keys = 2 ** len(first_zeros)
-        if not remaining_zeros:
-            yield key, n_keys
-            return
-        unwrapped_key = expand_to_bit_array(key)
-        for value in xrange(n_sets):
-            generated_key = numpy.copy(unwrapped_key)
-            generated_key[remaining_zeros] = \
-                expand_to_bit_array(value)[-len(remaining_zeros):]
-            yield compress_from_bit_array(generated_key), n_keys
 
     def _allocate_fixed_keys_and_masks(self, keys_and_masks, fixed_mask):
         """ Allocate fixed keys and masks.
@@ -264,7 +223,7 @@ class MallocBasedRoutingInfoAllocator(ElementAllocatorAlgorithm):
                     "Cannot meet conflicting constraints")
 
             # Go through the mask sets and allocate
-            for key, n_keys in self._get_key_ranges(
+            for key, n_keys in get_key_ranges(
                     key_and_mask.key, key_and_mask.mask):
                 self._allocate_elements(key, n_keys)
 
@@ -305,7 +264,7 @@ class MallocBasedRoutingInfoAllocator(ElementAllocatorAlgorithm):
                 # Check if all the key ranges can be allocated
                 matched_all = True
                 index = 0
-                for (base_key, n_keys) in self._get_key_ranges(key, mask):
+                for (base_key, n_keys) in get_key_ranges(key, mask):
                     logger.debug("Finding slot for {}, n_keys={}",
                                  hex(base_key), n_keys)
                     index = self._find_slot(base_key, lo=index)
@@ -333,7 +292,7 @@ class MallocBasedRoutingInfoAllocator(ElementAllocatorAlgorithm):
         # If we found a working key and mask that can be assigned,
         # Allocate them
         if key_found is not None and mask_found is not None:
-            for (base_key, n_keys) in self._get_key_ranges(key_found, mask):
+            for (base_key, n_keys) in get_key_ranges(key_found, mask):
                 self._allocate_elements(base_key, n_keys)
 
             # If we get here, we can assign the keys to the edges
